Add ability to show individual messages, including MIME parts
The MIME parts are unpacked into a directory 'static,' which happens to
be what the web.py dev server looks for.  It would be nice if this were
changeable so that static files could be served by
Apache/lighty/whatever.

The lack of escaping should be scaring me more than it is.  Must fix
that before sharing this more broadly.

multipart/alternative should show only one part, probably?  And
text/plain should have better wrapping, maybe?

And the UI is ugly.  But this is now something I could almost use to
read my mail.